fix:re Issue 609

+++ contracts/handler/SwapHandlers/PancakeSwapHandler.sol
@@ -45,7 +45,7 @@ contract PancakeSwapHandler is ISwapHandler {
   ) public view returns (bytes memory data) {
     bytes memory path = abi.encodePacked(
       tokenIn, // Address of the input token
-      fee, // Pool fee (0.3%)
+      uint24(fee), // Pool fee (0.3%)
       tokenOut // Address of the output token
     );
